Root/requirements.txt
# requirements.txt
pytest-8.3.5